Phase out vanilla three.js import from 3dObject
This is the only place that used three.js before me on this branch. Therefore, it is not too late to use Threebox instead, and it includes three.js, so the duplicate (the whole package) must be phased out.

GLTFLoader is collateral to this, and as such Threebox becomes contagious.