Help Threebox initialize successfully by adding accidentally required option
This is missing if not inserted by the user due to a wrong scope in the Threebox.init function, only for this one option.

The next commit could be one related to how this is edited inside Threebox to actually fix it, except Threebox is in the untracked node_modules. Either way, a GitHub issue sounds sensible.